eng-27: Rename imports
